Make Discover extension registrations async
