+++ src/plugins/discover/public/plugin.tsx
@@ -85,7 +85,9 @@ export interface RegisterExtensionsContext {
   stateContainer: DiscoverStateContainer;
 }
 
-export type RegisterExtensions = (options: RegisterExtensionsContext) => void | (() => void);
+export type RegisterExtensions = (
+  options: RegisterExtensionsContext
+) => void | (() => void) | Promise<void | (() => void)>;
 
 /**
  * @public
